Upgraded jMock 2.5.1 --> 2.12.0
Switched to ByteBuddy for the imposteriser.  I was using `cglib` before,
but it is not compatible with Java 16+ and there are no signs it will
get fixed any time soon.  I can now compile and run tests using Java 19.

https://github.com/cglib/cglib/issues/191

The JUnit 4 integration needed a different `Mockery`, for some reason.

I had to add a no-arg constructor to `ConstantPool` so that jMock could
mock it with the new ByteBuddy imposteriser.

ByteBuddy has a lot of difficulty imposterizing some classes.  One test
was trying to mock `PrintWriter` and I had to rewrite it to validate the
output using a wrapped `StringWriter` instead.