Document ClassCohesion CLI tool
I'm not sure LCOM4 is really that useful.  And my implementation is
limited by the impossibility to track dependencies on primitive
constants.

I refactor =ClassCohesion= to use the Streams API.  I could do better
when generating JSON.  There is still a `StringBuilder` in the middle
of it that I wouldn't mind getting rid of.  Maybe I could make that JSON
output prettier, while I'm at it.  I could manage with `jq` for now.

The tool is only on the CLI.  I will have to extract printers if I want
to have a matching Ant task.