Use Streams API in JSPs
I use `switch` expressions, which are a Java 14 feature.  If you deploy
Dependency Finder in Apache Tomcat, you need to let the JSP compiler
know to use advanced Java features.  I had to adjust the
`compilerSourceVM` and `compilerTargetVM` initialization parameters for
the `jsp` servlet in the `$CATALINA_HOME/conf/web.xml` file and set them
to `17` or larger.