Skip dependencies on primitive arrays
`StackMapTable_attribute` inadvertently introduced dependencies on
arrays of primitives.  For example, a local variable might be `byte[]`,
with the descriptor `[B`.  A stack frame would reference a `Class_info`
in the constant pool with that descriptor.  `CodeDependencyCollector`
woud try to get a class name from this descriptor and get `B[]` by
mistake.

In the process of fixing this bug, I refactored `DescriptorHelper` to
use a custom `Spliterator` from the Streams API instead of implementing
a custom `Enumeration`.