Move logic to build test archives from Ant to Gradle
Moved the Ant logic in `integration-tests/jarjardiff/build.xml` to the
`:integration-tests:jarjardiff:old` subproject.  I managed to create the
archives in the `old` subproject, but I'm not quite sure on the best way
to move them back up to either the `jarjardiff` subproject or the
`integration-tests` subproject.  For now, the tests that use these
archives use explicit paths into the `old` subproject's `build/` folder.

Now, I can run `./gradlew check` from a clean folder, and it takes care
of everything.