Add requested percentiles to StatisticalMeasurement
`StatisticalMeasurement.sum` used to default to `0.0` when the
measurement was empty, whereas the other values are `NaN`.  I changed it
to be inline with the other values.

Refactored `CVSPrinter` to use the Streams API to build text, including
requested percentiles.

Refactored `MeasurementBase.isInRange()` to simplify the logic.  It
parses thresholds once, when it creates the measurement's descriptor,
instead of doing it everytime `isInRange()` is called.

Also, minor miscellaneous clean ups.