Use isSynthetic() helper instead of only attribute
It is more reliable to use the `isSynthetic()` helper, which can check
both the presence of a `Synthetic_attribute` or the `ACC_SYNTHETIC`
access flag.

I updated `ClassMetrics` to cover the latest kinds of classes, fields,
methods, and inner classes.

I refactored the print methods in `ClassMetrics` to remove duplication.