Refactor finders to use predicates
Refactor `Classfile` to replace:

- `getField(name)`
- `locateField(name)`
- `getMethod(signature)`
- `locateMethod(signature)`

with equivalent methods that take a `Predicate` instead.  This way, the
caller can decide if they want to match on (name, signature, unique
name, etc.)

One downside is that the unit tests with mocks do not validate that the
predicate matches on the expected name.  I would have to re-write them
to use `Action` in order to exercise the predicate, and I don't feel
like it is worth the effort.