Refactor SymbolGatherer
Merge `Collector` (and `CollectorBase`) into it.  Replace collection
with a stream.  Sort and de-dupe in `ListSymbols` instead.  Moved some
of the navigation for local variables and inner classes to strategy
methods to help with testing (I can mock the strategy in tests).

`CodeDependencyCollector` is not really a `Collector`.  It does not
collect strings and no one is calling its `getCollection()` method.  I
made it inherit from `VisitorBase` instead, which is what it really is.