Pass in build version number from command-line
Instead of hard-coding the version number in Ant's `build.xml`, we can
pass it on the command-line using the `-Drelease=` option.  It defaults
to `SNAPSHOT` for casual builds.

In Gradle, we rely on `project.version` instead of some environment
variable.  We can set with the `-Pversion=` option.

I used to use the `build.tomcat.xml` Ant configuration to run Tomcat
locally and deploy the WAR file to it.  I don't use this setup anymore.
I run Tomcat in a Docker container instead, as configured by the
`Dockerfile`.